refactor: move expo handlers into separate files
