+++ package/expo-package/src/handlers/index.ts
@@ -1,0 +1,12 @@
+export * from './compressImage';
+export * from './deleteFile';
+export * from './getLocalAssetUri';
+export * from './getPhotos';
+export * from './NetInfo';
+export * from './pickDocument';
+export * from './saveFile';
+export * from './shareImage';
+export * from './Sound';
+export * from './takePhoto';
+export * from './triggerHaptic';
+export * from './Video';
